requirements.txt
clip==1.0
debugpy==1.8.2
einops==0.8.1
fastapi==0.115.12
lightning==2.3.3
matplotlib==3.10.3
mmcv_full==1.7.2
numpy==1.23.5
opencv_python==4.10.0.84
packaging==25.0
Pillow==11.2.1
scipy==1.15.3
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.14.0
uvicorn==0.34.3
